Restructure route to not use uid (#149)
* Restructure route to not use uid

* chapter member view

---------
