--- src/app/private/[uid]/admin/layout.tsx
@@ -1,52 +1,0 @@
-"use client";
-
-import { CollapsibleSidebarContainer } from "@components/container";
-import {
-  faEnvelope,
-  faHome,
-  faHouseLock,
-  faUser,
-} from "@fortawesome/free-solid-svg-icons";
-import React, { useContext } from "react";
-import { UserContext } from "src/context/UserProvider";
-
-interface IAdminLayout {
-  children: React.ReactNode;
-}
-
-const AdminLayout = ({ children }: IAdminLayout) => {
-  const { user } = useContext(UserContext);
-  const buttons = React.useMemo(
-    () => [
-      {
-        name: "Home",
-        link: `/private/${user.id}/admin/home`,
-        icon: faHome,
-      },
-      {
-        name: "Pending",
-        link: `/private/${user.id}/admin/pending-chapters`,
-        icon: faHouseLock,
-      },
-      {
-        name: "E-list",
-        link: `/private/${user.id}/admin/elist`,
-        icon: faEnvelope,
-      },
-      {
-        name: "Profile",
-        link: `/private/${user.id}/admin/edit-profile`,
-        icon: faUser,
-      },
-    ],
-    [user.id]
-  );
-
-  return (
-    <CollapsibleSidebarContainer buttons={buttons}>
-      {children}
-    </CollapsibleSidebarContainer>
-  );
-};
-
-export default AdminLayout;
