--- src/app/private/[uid]/chapter-leader/layout.tsx
@@ -1,54 +1,0 @@
-"use client";
-
-import Sidebar, { ISideBar } from "@components/Sidebar";
-import { CollapsibleSidebarContainer } from "@components/container";
-import {
-  faHome,
-  faUser,
-  faUserGroup,
-  faUserPlus,
-} from "@fortawesome/free-solid-svg-icons";
-import React, { useContext } from "react";
-import { UserContext } from "src/context/UserProvider";
-
-interface IUserLayout {
-  children: React.ReactNode;
-}
-
-const UserLayout = ({ children }: IUserLayout) => {
-  const { user } = useContext(UserContext);
-  const buttons: ISideBar["buttons"] = React.useMemo(
-    () => [
-      {
-        name: "Home",
-        link: `/private/${user.id}/chapter-leader/home`,
-        icon: faHome,
-      },
-      {
-        name: "Members",
-        link: `/private/${user.id}/chapter-leader/users`,
-        icon: faUserGroup,
-      },
-      // @TODO(nickbar01234) - Fix icon
-      {
-        name: "Seniors",
-        link: `/private/${user.id}/chapter-leader/seniors`,
-        icon: faUserGroup,
-      },
-      {
-        name: "Profile",
-        link: `/private/${user.id}/chapter-leader/edit-profile`,
-        icon: faUser,
-      },
-    ],
-    [user.id]
-  );
-
-  return (
-    <CollapsibleSidebarContainer buttons={buttons}>
-      {children}
-    </CollapsibleSidebarContainer>
-  );
-};
-
-export default UserLayout;
