+++ src/app/private/admin/edit-profile/page.tsx
